school_management_system/school_management_system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
APScheduler==3.10.4
redis==4.6.0
celery==5.3.1
python-dateutil==2.8.2
twilio==8.5.0
stripe==7.8.0
paypalrestsdk==1.13.1
zoomus==1.1.5
google-api-python-client==2.95.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
scikit-learn==1.3.0
numpy==1.24.3
nltk==3.8.1
transformers==4.31.0
torch==2.0.1
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1